Migrates SystemAdmin component routing to new structure
