+++ package.json
@@ -153,6 +153,7 @@
     "mockdate": "^3.0.5",
     "node-sass": "^6.0.1",
     "prettier": "2.4.1",
+    "ra-data-fakerest": "^3.18.3",
     "react-scripts": "^4.0.3",
     "react-select-event": "^5.3.0",
     "react-test-renderer": "^17.0.1",
