+++ src/scenes/SystemAdmin/AccessCodes/AccessCodeList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Datagrid, Filter, List, SelectInput, TextField, TextInput } from 'react-admin';
 import AdminPagination from 'scenes/SystemAdmin/shared/AdminPagination';
-import styles from 'scenes/SystemAdmin/Home.module.scss';
+import styles from 'pages/Admin/Home.module.scss';
 import { SHIPMENT_OPTIONS } from 'shared/constants';
 
 const defaultSort = { field: 'code', order: 'DESC' };
