+++ src/scenes/SystemAdmin/Notifications/NotificationList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { List, Datagrid, TextField, Filter, TextInput, DateField } from 'react-admin';
 import AdminPagination from 'scenes/SystemAdmin/shared/AdminPagination';
-import styles from 'scenes/SystemAdmin/Home.module.scss';
+import styles from 'pages/Admin/Home.module.scss';
 
 const defaultSort = { field: 'service_member_id', order: 'ASC' };
