+++ src/scenes/SystemAdmin/TSPPs/TSPPList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Datagrid, List, TextField, DateField, Filter, TextInput } from 'react-admin';
 import AdminPagination from 'scenes/SystemAdmin/shared/AdminPagination';
-import styles from 'scenes/SystemAdmin/Home.module.scss';
+import styles from 'pages/Admin/Home.module.scss';
 
 const defaultSort = { field: 'performance_period_start', order: 'DESC' };
